feat(training): training convenience features for ZEB-117 bake-off
- Add tiny_ffn_expanded to --config choices (Model β, PR #231 config
  was missing from CLI dispatch)
- Add --resume-from flag for checkpoint resume/fine-tuning
- Fix bf16 dtype mismatch in contrastive_loss() — torch.full() needs
  explicit dtype under autocast to match sim_orig/sim_proj dtypes
- Use C xxhash (xxhash.xxh64_intdigest) in lookup_batch_projected()
  and lookup_from_keys() with pure-Python fallback (~6x faster)
- Vectorized scatter-add replaces Python loop in both lookup methods

+++ training/ct87/latent_projection.py
@@ -271,7 +271,7 @@ def contrastive_loss(
 
     # Scale projected similarities by temperature, mask diagonal
     logits = sim_proj / temperature
-    diag_mask = torch.full((n, n), 0.0, device=projected.device)
+    diag_mask = torch.full((n, n), 0.0, device=projected.device, dtype=sim_proj.dtype)
     diag_mask.fill_diagonal_(-1e9)
     logits = logits + diag_mask
 
@@ -281,7 +281,7 @@ def contrastive_loss(
         orig_data.fill_diagonal_(float("-inf"))  # exclude self
         _, top_indices = orig_data.topk(k, dim=1)
 
-        target_logits = torch.full((n, n), float("-inf"), device=projected.device)
+        target_logits = torch.full((n, n), float("-inf"), device=projected.device, dtype=sim_orig.dtype)
         target_logits.scatter_(1, top_indices, sim_orig.gather(1, top_indices))
         targets = torch.softmax(target_logits, dim=1)
